+++ src/Middleware/AccessLogMiddleware.php
@@ -6,28 +6,37 @@ namespace Shlinkio\Shlink\Common\Middleware;
 
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
-use Psr\Http\Message\UriInterface;
 use Psr\Http\Server\MiddlewareInterface;
 use Psr\Http\Server\RequestHandlerInterface;
 use Psr\Log\LoggerInterface;
 
+use function Functional\some;
 use function sprintf;
+use function str_ends_with;
 
 class AccessLogMiddleware implements MiddlewareInterface
 {
     public const LOGGER_SERVICE_NAME = self::class . '/Logger';
 
-    public function __construct(private readonly LoggerInterface $logger)
+    public function __construct(private readonly LoggerInterface $logger, private readonly array $ignoreList = [])
     {
     }
 
     public function process(ServerRequestInterface $request, RequestHandlerInterface $handler): ResponseInterface
     {
         $response = $handler->handle($request);
+        $path = $request->getUri()->getPath();
+        $query = $request->getUri()->getQuery();
+
+        // Do not log requests for ignored endpoints
+        if (some($this->ignoreList, fn (string $ignored) => str_ends_with($path, $ignored))) {
+            return $response;
+        }
+
         $this->logger->info(sprintf(
             '%s %s %s %s',
             $request->getMethod(),
-            $this->requestUri($request->getUri()),
+            $this->requestUri($path, $query),
             $response->getStatusCode(),
             $response->getHeaderLine('Content-Length'),
         ));
@@ -35,11 +44,8 @@ class AccessLogMiddleware implements MiddlewareInterface
         return $response;
     }
 
-    private function requestUri(UriInterface $uri): string
+    private function requestUri(string $path, string $query): string
     {
-        $query = $uri->getQuery();
-        $path = $uri->getPath();
-
         return $query === '' ? $path : sprintf('%s?%s', $path, $query);
     }
 }